+++ module2/2.2/main.c
@@ -7,26 +7,26 @@ void menu() {
         printf("%s", *main_menu);
         int choice = -1;
         scanf("%d", &choice);
-        int x, y;
+        float x, y;
         if(choice && choice >= 1 && choice <= 4) {
             printf("Введите первый аргумент: ");
-            scanf(" %d", &x);
+            scanf(" %f", &x);
             printf("Введите второй аргумент: ");
-            scanf(" %d", &y);
+            scanf(" %f", &y);
         }
         switch (choice) {
             case 1:
-                printf("%d + %d = %d\n", x, y, add(x, y));
+                printf("%.3f + %.3f = %.3f\n", x, y, add(x, y));
                 break;
             case 2:
-                printf("%d - %d = %d\n", x, y, sub(x, y));
+                printf("%.3f - %.3f = %.3f\n", x, y, sub(x, y));
                 break;
             case 3:
-                printf("%d * %d = %d\n", x, y, mul(x, y));
+                printf("%.3f * %.3f = %.3f\n", x, y, mul(x, y));
                 break;
             case 4:
                 if (y)
-                printf("%d / %d = %d\n", x, y, div(x, y));
+                printf("%.3f / %.3f = %f\n", x, y, div(x, y));
                 else printf("На 0 делить нельзя!!\n");
                 break;
             case 0:
